- Add basic persistent settings (Mute and show tips, can be set only with keyboard during gameplay) - Cutscene reading, without asset check - Add in-game tips and other messages support - Add title animation - Fix serious bug which caused possibility of early level finish or no level transition after breaking required bricks - Fix ball appearing on middle when ball split is collected and at least one ball has been stuck on paddle

+++ Assets/Scripts/LevelPersistentData.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using System.Runtime.Serialization.Formatters.Binary;
 
 public class LevelPersistentData
 {
@@ -28,7 +27,7 @@ public class LevelPersistentData
 				if (readMagicNumber == magicNumber)
 				{
 					LevelNum = saveReader.ReadInt32();
-					DateTime dateTime = new DateTime(saveReader.ReadInt32());
+					DateTime dateTime = new DateTime(saveReader.ReadInt64());
 					if (levelSetDateTime != dateTime)
 						return false;
 					else
